Adds support to set log level from outside the pod
Previously we defined ddLogLevel as a static variable in the prefix file.
This declares it to every file in the pod but it means that we're stuck with
the default (const) value even if the app setting is different.

AFAIK, there's no way we can directly use the app's ddLogLevel as it needs to
be defined when the framework is linked, so the only way to control this is to
turn that const into a variable, and provide a couple of accessor methods.