Take two on setting log level
The previous attempt failed when added to the app. Instead of importing one of
the pod headers in the PCH, I imported the logging header whenever it was
needed.

Another problem was that the WPStatsLogging.h header was exporting a ddLogLevel
which might conflict with an existing declaration, so I moved that to a private
header. Now WPStatsLogging.h is the public interface to get/set the log level
and Logging.h is the internal header to import from every file that needs to
call DDLog*.